Refactor to fix swift lint issues
