+++ Timetracker/Views/TaskList/TaskListViewController.swift
@@ -10,7 +10,7 @@ import Foundation
 import Cocoa
 import SwiftDate
 
-class TaskListViewController: TrackedViewController, NSTableViewDataSource, NSTableViewDelegate {
+class TaskListViewController: TrackedViewController, NSTableViewDataSource {
 
     enum TableColumns: Int {
         case headOfDevelopment = 0
@@ -104,9 +104,7 @@ class TaskListViewController: TrackedViewController, NSTableViewDataSource, NSTa
         } else {
             editingTask = tasks?[safe: tableView.clickedRow]
             performSegue(withIdentifier: "add_line", sender: self)
-
         }
-
     }
 
     func numberOfRows(in tableView: NSTableView) -> Int {
@@ -121,88 +119,6 @@ class TaskListViewController: TrackedViewController, NSTableViewDataSource, NSTa
 
     func tableView(_ tableView: NSTableView, objectValueFor tableColumn: NSTableColumn?, row: Int) -> Any? {
         return tasks?[safe: row] ?? TaskProviderManager.instance.runningTask
-    }
-
-    // MARK: - TableViewDelegate callbacks
-
-    func handlePossibleCorruptContent(_ text: String?) -> String {
-        if let sureText = text {
-            return sureText
-        } else {
-            self.contentCorrupted = true
-            return "⚠️"
-        }
-    }
-
-    func buildContent(forTask task: Task, inProject project: Project?,
-                      tableColumn: NSTableColumn, inTable tableView: NSTableView) -> String {
-
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd HH:mm"
-
-        var text: String?
-
-        if tableColumn == tableView.tableColumns[TableColumns.headOfDevelopment.rawValue] { // HoD column
-            text = handlePossibleCorruptContent(project?.client?.headOfDevelopment?.name)
-
-        } else if tableColumn == tableView.tableColumns[TableColumns.client.rawValue] { // Client column
-            text = handlePossibleCorruptContent(project?.client?.name)
-
-        } else if tableColumn == tableView.tableColumns[TableColumns.project.rawValue] { // Project column
-            text = handlePossibleCorruptContent(project?.name)
-
-        } else if tableColumn == tableView.tableColumns[TableColumns.task.rawValue] { // Task name column
-            text = task.title
-
-        } else if tableColumn == tableView.tableColumns[TableColumns.startTime.rawValue] { // Task start column
-            if let startTime = task.startTime {
-                text = formatter.string(from: startTime as Date)
-            }
-
-        } else if tableColumn == tableView.tableColumns[TableColumns.endTime.rawValue] { // Task end column
-            if let endTime = task.endTime {
-                text = formatter.string(from: endTime as Date)
-            }
-
-        } else if tableColumn == tableView.tableColumns[TableColumns.accumulated.rawValue] { // Accumulated column
-            if let endTime = task.endTime {
-                let seconds = endTime.timeIntervalSince(task.startTime!)
-                text = seconds.toProperString()
-
-            } else {
-                let seconds = Date().timeIntervalSince(task.startTime!)
-                text = seconds.toProperString()
-            }
-        }
-
-        return text ?? ""
-
-    }
-
-    func tableView(_ tableView: NSTableView, viewFor tableColumn: NSTableColumn?, row: Int) -> NSView? {
-
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd HH:mm"
-
-        var task = tasks?[safe: row]
-        var project = task?.project
-
-        if task == nil && TaskProviderManager.instance.isTaskRunning {
-            task = TaskProviderManager.instance.runningTask
-            project = TaskProviderManager.instance.projectOfRunningTask
-        }
-
-        let text = buildContent(forTask: task!, inProject: project, tableColumn: tableColumn!, inTable: tableView)
-
-        let identifier: String = "hod_cell"
-
-        if let cell = tableView.makeView(withIdentifier: convertToNSUserInterfaceItemIdentifier(identifier),
-                                         owner: nil) as? NSTableCellView {
-            cell.textField?.stringValue = text
-            return cell
-        }
-
-        return nil
     }
 
     // MARK: - SegmentedControl action
@@ -260,7 +176,8 @@ class TaskListViewController: TrackedViewController, NSTableViewDataSource, NSTa
 
     @IBAction func exportClicked(_ sender: AnyObject) {
         guard !self.contentCorrupted else {
-            showError("There are tasks that do not seem to be properly configured. Please check and fix them.", because: "task-list-corrupted-content")
+            showError("There are tasks that do not seem to be properly configured. Please check and fix them.",
+                      because: "task-list-corrupted-content")
             return
         }
 
@@ -398,9 +315,4 @@ class TaskListViewController: TrackedViewController, NSTableViewDataSource, NSTa
     }
     //swiftlint:enable shorthand_operator
 
-}
-
-// Helper function inserted by Swift 4.2 migrator.
-private func convertToNSUserInterfaceItemIdentifier(_ input: String) -> NSUserInterfaceItemIdentifier {
-	return NSUserInterfaceItemIdentifier(rawValue: input)
 }
